+++ test/integration/pairs_test.lua
@@ -806,7 +806,7 @@ pgroup.test_opts_not_damaged = function(g)
     local after = {"Mary", 46, 2}
 
     local pairs_opts = {
-        timeout = 1, bucket_id = 1161,
+        timeout = 1, request_timeout = 1, bucket_id = 1161,
         batch_size = 105, first = 2, after = after,
         fields = fields, mode = 'write', prefer_replica = false,
         balance = false, force_map_call = false, use_tomap = true,