+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+* Support for `vshard`'s `request_timeout` parameter for calls with `mode = 'read'`
+
+### Changed
+
+### Fixed
+
 ## [1.6.0] - 09-09-25
 
 ### Changed